api/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
redis==5.0.1
mlflow==2.8.1
numpy==1.24.3
Pillow==10.1.0
kafka-python==2.0.2
python-jose[cryptography]==3.3.0
python-dotenv==1.0.0
